broker: move module handlers into modhash.c
Problem: the message handlers for managing modules are in
the huge broker.c, which makes them tricky to follow.

Move module message handlers over to modhash.c and register
them under the "module" service name:
  broker.insmod -> module.load
  broker.rmmod -> module.remove
  broker.lsmod -> module.list
  broker.module-status -> module.status
  broker.module-debug -> module.debug

Other than the service name change and the source relocation,
there are no substantive code changes in this commit.

Update users and tests.
